requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
tortoise-orm>=0.20.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
coverage>=7.0.0
httpx>=0.27.0
email-validator>=2.0.0

